Collapse column selections into multiselect
