+++ javascript/react/src/selectStepComponent.ts
@@ -25,7 +25,6 @@ import {
 	Rename,
 	Rollup,
 	Sample,
-	Select,
 	SetOperation,
 	Spread,
 	NoParameters,
@@ -59,7 +58,7 @@ const verb: Record<string, React.FC<StepComponentProps>> = {
 	rename: Rename,
 	rollup: Rollup,
 	sample: Sample,
-	select: Select,
+	select: ColumnListOperation,
 	spread: Spread,
 	unfold: Unfold,
 	ungroup: NoParameters,
