+++ javascript/react/src/selectStepDescription.ts
@@ -25,7 +25,6 @@ import {
 	RenameDescription,
 	RollupDescription,
 	SampleDescription,
-	SelectDescription,
 	SpreadDescription,
 	SetOperationDescription,
 	NoParametersDescription,
@@ -58,7 +57,7 @@ const descriptions: Record<string, React.FC<StepDescriptionProps>> = {
 	rename: RenameDescription,
 	rollup: RollupDescription,
 	sample: SampleDescription,
-	select: SelectDescription,
+	select: ColumnListOperationDescription,
 	spread: SpreadDescription,
 	unfold: UnfoldDescription,
 	ungroup: NoParametersDescription,
