+++ javascript/react/src/verbs/Fold/Fold.tsx
@@ -2,24 +2,14 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type {
-	ColumnListStep,
-	FoldStep,
-	Step,
-} from '@data-wrangling-components/core'
-import { ActionButton, Label, TextField } from '@fluentui/react'
-import type ColumnTable from 'arquero/dist/types/table/column-table'
-import set from 'lodash-es/set.js'
-import { memo, useCallback, useMemo } from 'react'
+import type { FoldStep } from '@data-wrangling-components/core'
+import { TextField } from '@fluentui/react'
+import { memo, useMemo } from 'react'
 import styled from 'styled-components'
-import {
-	LeftAlignedRow,
-	useHandleTextfieldChange,
-	useLoadTable,
-} from '../../common'
-import { ColumnInstruction } from '../../controls'
+import { LeftAlignedRow, useHandleTextfieldChange } from '../../common'
 import { dropdownStyles } from '../../controls/styles.js'
 import type { StepComponentProps } from '../../types'
+import { ColumnListInputs } from '../shared/index.js'
 
 /**
  * Provides inputs for a step that needs lists of columns.
@@ -32,21 +22,6 @@ export const Fold: React.FC<StepComponentProps> = memo(function Fold({
 	input,
 }) {
 	const internal = useMemo(() => step as FoldStep, [step])
-
-	const tbl = useLoadTable(input || step.input, table, store)
-
-	const columns = useColumns(internal, tbl, onChange)
-
-	const handleButtonClick = useCallback(() => {
-		onChange &&
-			onChange({
-				...internal,
-				args: {
-					...internal.args,
-					columns: [...internal.args.columns, first(tbl)],
-				},
-			})
-	}, [internal, tbl, onChange])
 
 	const handleToChange = useHandleTextfieldChange(
 		internal,
@@ -61,16 +36,15 @@ export const Fold: React.FC<StepComponentProps> = memo(function Fold({
 
 	return (
 		<Container>
-			<Label>Columns</Label>
-			{columns}
-			<ActionButton
-				onClick={handleButtonClick}
-				iconProps={{ iconName: 'Add' }}
-				disabled={!tbl}
-			>
-				Add column
-			</ActionButton>
-
+			<LeftAlignedRow>
+				<ColumnListInputs
+					step={step}
+					store={store}
+					table={table}
+					onChange={onChange}
+					input={input}
+				/>
+			</LeftAlignedRow>
 			<LeftAlignedRow>
 				<TextField
 					required
@@ -94,43 +68,6 @@ export const Fold: React.FC<StepComponentProps> = memo(function Fold({
 		</Container>
 	)
 })
-
-function first(table?: ColumnTable): string {
-	return table?.columnNames()[0] as string
-}
-
-function useColumns(
-	step: ColumnListStep,
-	table?: ColumnTable,
-	onChange?: (step: Step) => void,
-) {
-	return useMemo(() => {
-		return (step.args.columns || []).map((column: string, index: number) => {
-			const handleColumnChange = (col: string) => {
-				const update = { ...step }
-				set(update, `args.columns[${index}]`, col)
-				onChange && onChange(update)
-			}
-
-			const handleDeleteClick = () => {
-				const update = { ...step }
-				update.args.columns.splice(index, 1)
-				onChange && onChange(update)
-			}
-
-			return (
-				<LeftAlignedRow key={`column-list-${column}-${index}`}>
-					<ColumnInstruction
-						table={table}
-						column={column}
-						onChange={handleColumnChange}
-						onDelete={handleDeleteClick}
-					/>
-				</LeftAlignedRow>
-			)
-		})
-	}, [step, table, onChange])
-}
 
 const Container = styled.div`
 	display: flex;
