+++ javascript/react/src/verbs/index.ts
@@ -26,6 +26,5 @@ export * from './Recode/index.js'
 export * from './Rename/index.js'
 export * from './Rollup/index.js'
 export * from './Sample/index.js'
-export * from './Select/index.js'
 export * from './NoParameters/index.js'
 export * from './Window/index.js'
